Added new header for visiter and added setting for optimization
